hwc/nsc: add conversion yv12 planar to ycrcb packed.
nsc cannot handle planar format natively, therefore when
such is encountered we need an extra step to do a conversion
from planar to packed ahead of submitting the surface for
composition to nsc.

conversion code is lifted from nexus soft video decoder
example.

todo: we may want to consider doing the conversion async
between prepare and set.  but for now, let's see how this holds
up.

refs #SWANDROID-1050

+++ bcm_platform/libgralloc/gralloc.cpp
@@ -538,7 +538,7 @@ gralloc_alloc_buffer(alloc_device_t* dev,
       pSharedData->planes[EXTRA_PLANE].width = w;
       pSharedData->planes[EXTRA_PLANE].height = h;
       pSharedData->planes[EXTRA_PLANE].bpp = bpp;
-      pSharedData->planes[EXTRA_PLANE].format = nxFormat;
+      pSharedData->planes[EXTRA_PLANE].format = (int)nxFormat;
       ret = ioctl(fd3, NX_ASHMEM_SET_SIZE, extra_size);
       if (ret >= 0) {
          pSharedData->planes[EXTRA_PLANE].physAddr =
